+++ backend/src/main/java/com/woowacourse/zzimkkong/infrastructure/auth/AuthorizationExtractor.java
@@ -1,4 +1,4 @@
-package com.woowacourse.zzimkkong.infrastructure;
+package com.woowacourse.zzimkkong.infrastructure.auth;
 
 import com.woowacourse.zzimkkong.exception.authorization.AuthorizationHeaderUninvolvedException;
