+++ backend/src/main/java/com/woowacourse/zzimkkong/infrastructure/sharingId/SharingIdGenerator.java
@@ -1,4 +1,4 @@
-package com.woowacourse.zzimkkong.infrastructure;
+package com.woowacourse.zzimkkong.infrastructure.sharingId;
 
 import com.woowacourse.zzimkkong.domain.Map;
 import com.woowacourse.zzimkkong.exception.infrastructure.DecodingException;
