+++ backend/src/main/java/com/woowacourse/zzimkkong/infrastructure/sharingId/Transcoder.java
@@ -1,4 +1,4 @@
-package com.woowacourse.zzimkkong.infrastructure;
+package com.woowacourse.zzimkkong.infrastructure.sharingId;
 
 public interface Transcoder {
     String encode(String input);
