+++ backend/src/main/java/com/woowacourse/zzimkkong/infrastructure/thumbnail/BatikConverter.java
@@ -1,4 +1,4 @@
-package com.woowacourse.zzimkkong.infrastructure;
+package com.woowacourse.zzimkkong.infrastructure.thumbnail;
 
 import com.woowacourse.zzimkkong.exception.infrastructure.SvgToPngConvertException;
 import org.apache.batik.transcoder.TranscoderException;
