+++ backend/src/main/java/com/woowacourse/zzimkkong/infrastructure/thumbnail/S3Uploader.java
@@ -1,4 +1,4 @@
-package com.woowacourse.zzimkkong.infrastructure;
+package com.woowacourse.zzimkkong.infrastructure.thumbnail;
 
 import com.amazonaws.AmazonClientException;
 import com.amazonaws.services.s3.AmazonS3;
