+++ backend/src/main/java/com/woowacourse/zzimkkong/infrastructure/thumbnail/SvgConverter.java
@@ -1,4 +1,4 @@
-package com.woowacourse.zzimkkong.infrastructure;
+package com.woowacourse.zzimkkong.infrastructure.thumbnail;
 
 import java.io.File;
