+++ backend/src/test/java/com/woowacourse/zzimkkong/infrastructure/JwtUtilsTest.java
@@ -4,6 +4,7 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.woowacourse.zzimkkong.exception.authorization.InvalidTokenException;
 import com.woowacourse.zzimkkong.exception.authorization.TokenExpiredException;
+import com.woowacourse.zzimkkong.infrastructure.auth.JwtUtils;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
