+++ backend/src/test/java/com/woowacourse/zzimkkong/service/AuthServiceTest.java
@@ -9,7 +9,7 @@ import com.woowacourse.zzimkkong.exception.authorization.OauthProviderMismatchEx
 import com.woowacourse.zzimkkong.exception.member.NoSuchMemberException;
 import com.woowacourse.zzimkkong.exception.member.NoSuchOAuthMemberException;
 import com.woowacourse.zzimkkong.exception.member.PasswordMismatchException;
-import com.woowacourse.zzimkkong.infrastructure.JwtUtils;
+import com.woowacourse.zzimkkong.infrastructure.auth.JwtUtils;
 import com.woowacourse.zzimkkong.infrastructure.oauth.OauthHandler;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -20,9 +20,7 @@ import org.junit.jupiter.params.provider.EnumSource;
 import org.junit.jupiter.params.provider.MethodSource;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.security.core.parameters.P;
 
-import java.util.Arrays;
 import java.util.Optional;
 import java.util.stream.Stream;
